Add cart functionality with CreateCart command, validation, and mapping updates

+++ src/Application/Features/Carts/Dtos/SaleProductDto.cs
@@ -2,6 +2,6 @@ namespace DeveloperStore.Application.Features.Carts.Dtos;
 
 public class SaleProductDto
 {
-    public int ProductId { get; set; }
+    public long ProductId { get; set; }
     public int Quantity { get; set; }
 }
